Move client to subfolder, create post class
