+++ api-sdk/src/client/index.ts
@@ -1,0 +1,3 @@
+import Client from "./client";
+
+export default Client;
